+++ mapsforge-map-android/src/main/java/org/mapsforge/map/android/graphics/AndroidPaint.java
@@ -1,7 +1,7 @@
 /*
  * Copyright 2010, 2011, 2012, 2013 mapsforge.org
  * Copyright 2014 Ludwig M Brinckmann
- * Copyright 2015-2016 devemux86
+ * Copyright 2015-2017 devemux86
  *
  * This program is free software: you can redistribute it and/or modify it under the
  * terms of the GNU Lesser General Public License as published by the Free Software
@@ -142,6 +142,11 @@ class AndroidPaint implements Paint {
     @Override
     public int getColor() {
         return this.paint.getColor();
+    }
+
+    @Override
+    public float getStrokeWidth() {
+        return paint.getStrokeWidth();
     }
 
     @Override